Fixed pagination + Add tags to all views
