--- FA.JustBlog/Views/Home/_PartialAboutCard.cshtml
@@ -1,34 +1,0 @@
-﻿@model IEnumerable<FA.JustBlog.Models.TagViewModel>
-
-<div class="position-relative text-center">
-	<img src="~/img/about-bg.jpg" style="width: 100%; opacity: 0.5" asp-append-version=true>
-	<div class="container text-black position-absolute px-4 px-lg-5" style="top: 50%; left: 50%; transform: translate(-50%, -50%)">
-		<div class=" text-center gx-4 gx-lg-5">
-
-			<div class="page-heading" style="z-index: 2">
-				<h1>About Me</h1>
-				<span>This is what I do.</span>
-			</div>
-		</div>
-	</div>
-</div>
-
-<div style="text-align:justify">
-	<p>
-		Lorem ipsum dolor sit amet, consectetur adipisicing elit.
-		Saepe nostrum ullam eveniet pariatur voluptates odit, fuga atque ea nobis sit soluta odio,
-		adipisci quas excepturi maxime quae totam ducimus consectetur?
-	</p>
-</div>
-
-<div style="position-relative text-center">
-	@if (Model != null)
-	{
-		foreach (var item in Model)
-		{
-			<span class="tag-box">
-				@Html.DisplayFor(modelItem => item.Name)
-			</span>
-		}
-	}
-</div>
