+++ FA.JustBlog/Views/Post/LatestPost.cshtml
@@ -54,7 +54,7 @@
 <div class="pagination">
     @if (Model.HasPreviousPage)
     {
-        <a href="@Url.Action("Index", new { page = Model.PageNumber - 1 })">Previous</a>
+        <a href="@Url.Action("LatestPost", new { page = Model.PageNumber - 1 })">Previous</a>
     }
 
     @for (int i = 1; i <= Model.PageCount; i++)
@@ -65,13 +65,13 @@
         }
         else
         {
-            <a href="@Url.Action("Index", new { page = i })">@i</a>
+            <a href="@Url.Action("LatestPost", new { page = i })">@i</a>
         }
     }
 
     @if (Model.HasNextPage)
     {
-        <a href="@Url.Action("Index", new { page = Model.PageNumber + 1 })">Next</a>
+        <a href="@Url.Action("LatestPost", new { page = Model.PageNumber + 1 })">Next</a>
     }
 </div>
